requirements.txt
anthropic
tavily-python
requests
beautifulsoup4
jinja2
pyyaml
